// src/models/GumastaFailedPaymentResponse.model.ts
export type GumastaFailedPaymentListResponse ={
    _id: string
    propritorName: string
    adhaarNumber: string
    mobileNumber: string
    email: string
    firmName: string
    firmAddress: string
    propritorPhotoUrl: string
    adhaarPhotoUrl: string
    shopOfficePhotoUrl: string
    addressProofPhotoUrl: string
    otherDocuments: any[] 
    state: string
    district: string
    distributorId: string
    distributorCode: string
    txnObjectId: string
    txnId: string
    uniqueTransactionId: string
    srn: string
    appliedFrom: string
    version: string
    acknowledgementNumber: string
    acknowledgementPdf: string
    status: string
    txnStatus: string
    comments: string
    totalPrice: number
    convinienceCharges: number
    basePrice: number
    baseCatAppied: string
    refundWalletAmountApplied: number
    rewardWalletAmountApplied: number
    rejectionReason: string
    remark: string
    appliedByUserCurrentAddress: string
    appliedByType: string
    appliedById: string
    appliedOnDate: string
    appliedByName: string
    appliedByNumber: string
    appliedAsType: string
    assignedToName: string
    assignedToId: string
    assignedOnDate: string
    assignedByName: string
    assignedById: string
    verifiedByName: string
    verifiedById: string
    verifiedOnDate: string
    rejectedByName: string
    rejectedById: string
    rejectedOnDate: string
    generatedByName: string
    generatedById: string
    generatedOnDate: string
    completedByName: string
    completedById: string
    completedOnDate: string
    cancelledByName: string
    cancelledById: string
    cancelledOnDate: string
    moveToPendingByName: string
    moveToPendingById: string
    moveToPendingOnDate: string
    IsAgreedToTermsAndConditions: boolean
    isDeleted: boolean
    createdAt: string
    updatedAt: string
    __v: number
  }